fix curriculum3
