+++ webapp/routes/api.php
@@ -18,5 +18,4 @@ use App\Http\Controllers\RemotePostController;
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
-
 Route::post('/posts', [RemotePostController::class, 'createNewPost']);
